// src/app/models/experience.model.ts
export interface Experience {
  id?: string;
  name?: string;
  type?: string;
  uid?: string;
  company?: string;
  location?: string;
  description?: string;
  hasDate?: boolean;
  startDate?: number;
  endDate?: number;
  current?: boolean;
  skills?: any[];
  link?: string;
  image?: string;
  links?: any[];
}